Get DLL configs running out of the box
Check in demo.vcxproj.user, which allows modification of the runtime
environment for the demo.

Remove demo.vcxproj.user from .gitignore.

Add glut/runtime to the path for all configs so they can find the DLL
there.

Delete the extra copy of glut32.dll from the demo directory.

Add chipmunk output directory to the path for the DLL configs which
gets them working (they weren't able to find the DLL on the search
path before).