Unified FFI by using non-static inlines for public functions
A separate translation unit is now responsible for generating exports,
as recommended by section 6.7.4 of the ISO C standard

This completes the task in *include/chipmunk/chipmunk_ffi.h*:

    // TODO: get rid of the reliance on static inlines.
    // They make a mess for FFIs.

Legacy function pointers are still supported, though the use of
-DCHIPMUNK_LEGACY_FFI. An additional option in cmake makes it easy to
enable or disable this feature. Eventually, it can be removed completely.