Update ChipmunkSpace.h
ChipmunkSpace has the read only property "locked", but in the implementation, the iVar _locked is never set. On the other hand, the implementation has an "isLocked" method (which actually checks the locking state of the associated chipmunk space).
So, think the readonly property on the header file should be "isLocked" too.